refactored to extract middleware functions and error handler to another file, refactored file structure under api
