+++ api/config/dbConfig.js
@@ -1,5 +1,5 @@
 const knex = require('knex');
-const knexConfig = require('../knexfile.js');
+const knexConfig = require('../../knexfile.js');
 const db = knex(knexConfig.development);
 
 module.exports = db;
